+++ lib/dicebag.rb
@@ -15,9 +15,24 @@ module DiceBag
     Normalize.call ast
   end
 
-  ### Main Syntatic Sugar Interface
+  ###
+  # Main Syntatic Sugar Interface Methodds
+  ###
+
   def self.roll(dstr)
     DiceBag::Roll.new(dstr).roll
+  end
+
+  def self.average(dstr)
+    DiceBag::Roll.new(dstr).average
+  end
+
+  def self.maximum(dstr)
+    DiceBag::Roll.new(dstr).maximum
+  end
+
+  def self.minimum(dstr)
+    DiceBag::Roll.new(dstr).minimum
   end
 
   # The default roll if one is not given.
@@ -27,14 +42,15 @@ module DiceBag
 end
 
 # Our sub-modules.
-require_relative './dicebag/normalize'
-require_relative './dicebag/roll_string'
-require_relative './dicebag/roll_part_string'
-require_relative './dicebag/parser'
-require_relative './dicebag/transform'
-require_relative './dicebag/simple_part'
-require_relative './dicebag/label_part'
-require_relative './dicebag/static_part'
-require_relative './dicebag/roll_part'
-require_relative './dicebag/roll'
-require_relative './dicebag/result'
+require_relative 'dicebag/normalize'
+require_relative 'dicebag/min_max_calc'
+require_relative 'dicebag/roll_string'
+require_relative 'dicebag/roll_part_string'
+require_relative 'dicebag/parser'
+require_relative 'dicebag/transform'
+require_relative 'dicebag/simple_part'
+require_relative 'dicebag/label_part'
+require_relative 'dicebag/static_part'
+require_relative 'dicebag/roll_part'
+require_relative 'dicebag/roll'
+require_relative 'dicebag/result'